Moved 2 components to own files for bundle size test
