+++ pages/accountLists/[accountListId]/reports/designationAccounts.page.tsx
@@ -9,7 +9,7 @@ import {
   MultiPageMenu,
   NavTypeEnum,
 } from 'src/components/Shared/MultiPageLayout/MultiPageMenu/MultiPageMenu';
-import { ReportPageWrapper } from 'src/components/Shared/styledComponents/styledComponents';
+import { ReportPageWrapper } from 'src/components/Shared/styledComponents/ReportPageWrapper';
 import { useAccountListId } from 'src/hooks/useAccountListId';
 import useGetAppSettings from 'src/hooks/useGetAppSettings';
