+++ src/components/Contacts/ContactDetails/ContactDetailsTab/Mailing/AddAddressModal/AddAddressModal.tsx
@@ -14,10 +14,10 @@ import { Formik } from 'formik';
 import { useSnackbar } from 'notistack';
 import { useTranslation } from 'react-i18next';
 import { AddressLocationSelect } from 'src/common/Selects/AddressLocationSelect';
+import { LoadingIndicator } from 'src/components/Shared/styledComponents/LoadingIndicator';
 import {
   ContactEditContainer,
   ContactInputWrapper,
-  LoadingIndicator,
 } from 'src/components/Shared/styledComponents/styledComponents';
 import {
   CancelButton,
