+++ src/components/Contacts/MassActions/Merge/MassActionsMergeModal.tsx
@@ -11,7 +11,7 @@ import {
 import { DateTime } from 'luxon';
 import { useSnackbar } from 'notistack';
 import { useTranslation } from 'react-i18next';
-import { LoadingIndicator } from 'src/components/Shared/styledComponents/styledComponents';
+import { LoadingIndicator } from 'src/components/Shared/styledComponents/LoadingIndicator';
 import {
   CancelButton,
   SubmitButton,
