+++ src/components/Shared/styledComponents/ReportPageWrapper.ts
@@ -1,0 +1,6 @@
+import { Box } from '@mui/material';
+import { styled } from '@mui/material/styles';
+
+export const ReportPageWrapper = styled(Box)(({ theme }) => ({
+  backgroundColor: theme.palette.common.white,
+}));
