+++ src/components/common/Modal/Confirmation/Confirmation.tsx
@@ -1,10 +1,8 @@
 import React, { ReactNode, useState } from 'react';
 import { Box, DialogActions, DialogContent } from '@mui/material';
 import { useTranslation } from 'react-i18next';
-import {
-  LoadingIndicator,
-  StyledDialogContentText,
-} from 'src/components/Shared/styledComponents/styledComponents';
+import { LoadingIndicator } from 'src/components/Shared/styledComponents/LoadingIndicator';
+import { StyledDialogContentText } from 'src/components/Shared/styledComponents/styledComponents';
 import {
   ActionButtonProps,
   CancelButton,
